trip_eld_backend/pyproject.toml
[project]
name = "trip-eld-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "geopy>=2.4.1",
    "markdown>=3.9",
    "openrouteservice>=2.3.3",
]
